Added new implementation

+++ kernel/include/common/memoryoperations.h
@@ -13,10 +13,10 @@ namespace CactusOS
         class MemoryOperations
         {
         public:
-            static void* memmove(void* dstptr, const void* srcptr, size_t size);
-            static int memcmp(const void* aptr, const void* bptr, size_t size);
-            static void* memset(void* bufptr, char value, size_t size);
-            static void* memcpy(void* dstptr, const void* srcptr, size_t size);
+            static void* memmove(void* dstptr, const void* srcptr, uint32_t size);
+            static int memcmp(const void* aptr, const void* bptr, uint32_t size);
+            static void* memset(void* bufptr, char value, uint32_t size);
+            static void* memcpy(void* dstptr, const void* srcptr, uint32_t size);
         };
     }
 }
