+++ kernel/src/common/memoryoperations.cpp
@@ -2,24 +2,24 @@
 
 using namespace CactusOS::common;
 
-void* MemoryOperations::memmove(void* dstptr, const void* srcptr, size_t size)
+void* MemoryOperations::memmove(void* dstptr, const void* srcptr, uint32_t size)
 {
     unsigned char* dst = (unsigned char*) dstptr;
 	const unsigned char* src = (const unsigned char*) srcptr;
 	if (dst < src) {
-		for (size_t i = 0; i < size; i++)
+		for (uint32_t i = 0; i < size; i++)
 			dst[i] = src[i];
 	} else {
-		for (size_t i = size; i != 0; i--)
+		for (uint32_t i = size; i != 0; i--)
 			dst[i-1] = src[i-1];
 	}
 	return dstptr;
 }
-int MemoryOperations::memcmp(const void* aptr, const void* bptr, size_t size)
+int MemoryOperations::memcmp(const void* aptr, const void* bptr, uint32_t size)
 {
     const unsigned char* a = (const unsigned char*) aptr;
 	const unsigned char* b = (const unsigned char*) bptr;
-	for (size_t i = 0; i < size; i++) {
+	for (uint32_t i = 0; i < size; i++) {
 		if (a[i] < b[i])
 			return -1;
 		else if (b[i] < a[i])
@@ -27,18 +27,18 @@ int MemoryOperations::memcmp(const void* aptr, const void* bptr, size_t size)
 	}
 	return 0;
 }
-void* MemoryOperations::memset(void* bufptr, char value, size_t size)
+void* MemoryOperations::memset(void* bufptr, char value, uint32_t size)
 {
     unsigned char* buf = (unsigned char*) bufptr;
-	for (size_t i = 0; i < size; i++)
+	for (uint32_t i = 0; i < size; i++)
 		buf[i] = (unsigned char) value;
 	return bufptr;
 }
-void* MemoryOperations::memcpy(void* dstptr, const void* srcptr, size_t size)
+void* MemoryOperations::memcpy(void* dstptr, const void* srcptr, uint32_t size)
 {
     unsigned char* dst = (unsigned char*) dstptr;
 	const unsigned char* src = (const unsigned char*) srcptr;
-	for (size_t i = 0; i < size; i++)
+	for (uint32_t i = 0; i < size; i++)
 		dst[i] = src[i];
 	return dstptr;
 }
